pyproject.toml
[project]
name = "multiagent-qfab-quality-assesment"
version = "0.1.0"
description = ""
authors = [
    {name = "kapedalex",email = "kapedalex@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "pytest (>=9.0.2,<10.0.0)",
    "numpy (>=2.4.3,<3.0.0)",
    "setuptools (>=82.0.1,<83.0.0)",
    "transformers (>=5.3.0,<6.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "torch (>=2.10.0,<3.0.0)",
    "together (>=2.4.0,<3.0.0)",
    "graphviz (>=0.21,<0.22)",
    "pygraphviz ==1.14",
    "ipykernel (>=7.2.0,<8.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "anthropic (>=0.84.0,<0.85.0)",
    "openai (>=2.26.0,<3.0.0)",
    "google-genai (>=0.8.0,<1.0.0)",
    "google-api-python-client (>=2.192.0,<3.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "mistralai (>=2.0.1,<3.0.0)",
    "datasets (>=4.7.0,<5.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "multiagent-qfab-quality-assesment"
version = "0.1.0"
description = ""
authors = ["kapedalex <kapedalex@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "common" }]

[tool.poetry.dependencies]
qbaf-library = { path = "Quantitative-Bipolar-Argumentation", develop = true }
